Remove unnecessary unboxing

+++ src/main/java/com/fasterxml/jackson/databind/util/TokenBuffer.java
@@ -449,11 +449,11 @@ public class TokenBuffer
                 {
                     Object n = segment.get(ptr);
                     if (n instanceof Double) {
-                        gen.writeNumber(((Double) n).doubleValue());
+                        gen.writeNumber((Double) n);
                     } else if (n instanceof BigDecimal) {
                         gen.writeNumber((BigDecimal) n);
                     } else if (n instanceof Float) {
-                        gen.writeNumber(((Float) n).floatValue());
+                        gen.writeNumber((Float) n);
                     } else if (n == null) {
                         gen.writeNull();
                     } else if (n instanceof String) {
